+++ docs/source/changelog.rst
@@ -3460,3 +3460,5 @@ Current C++/SQLite client, Python/SQLAlchemy server
   (Database revision 0066)
 
 - Fix Debian package Python dependencies. Would fail if Python 3.6 was not installed.
+
+- **New task:** :ref:`CPFT Research Preferences  <cpft_research_preferences>` (Database revision 0067).